+++ Rabbot/Commands/Combi.cs
@@ -35,17 +35,17 @@ namespace Rabbot.Commands
 
             using (swaightContext db = new swaightContext())
             {
-                if (db.Combi.Where(p => p.ServerId == (long)Context.Guild.Id && p.UserId == (long)Context.User.Id).Count() >= 5)
+                if (db.Combi.Where(p => p.ServerId == Context.Guild.Id && p.UserId == Context.User.Id).Count() >= 5)
                 {
                     await ReplyAsync($"{Context.User.Mention} du hast bereits 5 bestehende Combis.");
                     return;
                 }
-                if (db.Combi.Where(p => p.ServerId == (long)Context.Guild.Id && p.UserId == (long)user.Id).Count() >= 5)
+                if (db.Combi.Where(p => p.ServerId == Context.Guild.Id && p.UserId == user.Id).Count() >= 5)
                 {
                     await ReplyAsync($"{user.Mention} hat bereits 5 bestehende Combis.");
                     return;
                 }
-                if (db.Combi.Any(p => p.ServerId == (long)Context.Guild.Id && (p.UserId == (long)Context.User.Id && p.CombiUserId == (long)user.Id) || p.CombiUserId == (long)Context.User.Id && p.UserId == (long)user.Id))
+                if (db.Combi.Any(p => p.ServerId == Context.Guild.Id && (p.UserId == Context.User.Id && p.CombiUserId == user.Id) || p.CombiUserId == Context.User.Id && p.UserId == user.Id))
                 {
                     await ReplyAsync($"{Context.User.Mention} du hast bereits eine **bestehende** Combi oder Combi Anfrage mit {user.Mention}.");
                     return;
@@ -55,7 +55,7 @@ namespace Rabbot.Commands
                 embed.WithDescription($"Du hast Erfolgreich eine Combi Anfrage an {user.Mention} gestellt.");
                 embed.WithFooter("Die Anfrage kann über die Reactions angenommen oder abgelehnt werden.");
                 var msg = await ReplyAsync($"{user.Mention}", false, embed.Build());
-                await db.Combi.AddAsync(new Combi { ServerId = (long)Context.Guild.Id, UserId = (long)Context.User.Id, CombiUserId = (long)user.Id, Accepted = false, MessageId = (long)msg.Id, Date = DateTime.Now });
+                await db.Combi.AddAsync(new Combi { ServerId = Context.Guild.Id, UserId = Context.User.Id, CombiUserId = user.Id, Accepted = false, MessageId = msg.Id, Date = DateTime.Now });
                 await db.SaveChangesAsync();
                 var emoteAccept = new Emoji("✅");
                 var emoteDeny = new Emoji("⛔");
@@ -72,7 +72,7 @@ namespace Rabbot.Commands
         {
             using (swaightContext db = new swaightContext())
             {
-                var combis = db.Combi.Include(p => p.User).Include(p => p.CombiUser).Where(p => p.ServerId == (long)Context.Guild.Id && (p.UserId == (long)Context.User.Id || p.CombiUserId == (long)Context.User.Id));
+                var combis = db.Combi.Include(p => p.User).Include(p => p.CombiUser).Where(p => p.ServerId == Context.Guild.Id && (p.UserId == Context.User.Id || p.CombiUserId == Context.User.Id));
 
                 EmbedBuilder embed = new EmbedBuilder();
                 embed.WithTitle("Combis");
@@ -85,9 +85,9 @@ namespace Rabbot.Commands
                     else
                         status = "Ausstehend";
 
-                    if (combi.CombiUserId == (long)Context.User.Id)
+                    if (combi.CombiUserId == Context.User.Id)
                         embed.AddField($"[{count}] {combi.User.Name}", $"Status: {status}\nBesteht seit: {combi.Date.Value.ToString("dd.MM.yyyy")}");
-                    if (combi.UserId == (long)Context.User.Id)
+                    if (combi.UserId == Context.User.Id)
                         embed.AddField($"[{count}] {combi.CombiUser.Name}", $"Status: {status}\nBesteht seit: {combi.Date.Value.ToString("dd.MM.yyyy")}");
 
                     count++;
@@ -104,7 +104,7 @@ namespace Rabbot.Commands
         {
             using (swaightContext db = new swaightContext())
             {
-                var combi = db.Combi.Include(p => p.User).Include(p => p.CombiUser).Where(p => p.ServerId == (long)Context.Guild.Id && (p.UserId == (long)Context.User.Id || p.CombiUserId == (long)Context.User.Id)).Skip(id - 1)?.FirstOrDefault();
+                var combi = db.Combi.Include(p => p.User).Include(p => p.CombiUser).Where(p => p.ServerId == Context.Guild.Id && (p.UserId == Context.User.Id || p.CombiUserId == Context.User.Id)).Skip(id - 1)?.FirstOrDefault();
 
                 if (combi == null)
                 {
@@ -115,9 +115,9 @@ namespace Rabbot.Commands
                 db.Combi.Remove(combi);
                 await db.SaveChangesAsync();
 
-                if (combi.CombiUserId == (long)Context.User.Id)
+                if (combi.CombiUserId == Context.User.Id)
                     await ReplyAsync($"{Context.User.Mention} du hast erfolgreich die Combi mit **{combi.User.Name}** gelöscht.");
-                if (combi.UserId == (long)Context.User.Id)
+                if (combi.UserId == Context.User.Id)
                     await ReplyAsync($"{Context.User.Mention} du hast erfolgreich die Combi mit **{combi.CombiUser.Name}** gelöscht.");
             }
         }
@@ -130,7 +130,7 @@ namespace Rabbot.Commands
         {
             using (swaightContext db = new swaightContext())
             {
-                var combi = db.Combi.Include(p => p.User).Include(p => p.CombiUser).Where(p => p.ServerId == (long)Context.Guild.Id && (p.UserId == (long)Context.User.Id || p.CombiUserId == (long)Context.User.Id)).Skip(id - 1)?.FirstOrDefault();
+                var combi = db.Combi.Include(p => p.User).Include(p => p.CombiUser).Where(p => p.ServerId == Context.Guild.Id && (p.UserId == Context.User.Id || p.CombiUserId == Context.User.Id)).Skip(id - 1)?.FirstOrDefault();
 
                 if (combi == null)
                 {
@@ -146,9 +146,9 @@ namespace Rabbot.Commands
                 }
                 await db.SaveChangesAsync();
 
-                if (combi.CombiUserId == (long)Context.User.Id)
+                if (combi.CombiUserId == Context.User.Id)
                     await ReplyAsync($"{Context.User.Mention} du hast erfolgreich die Combi mit **{combi.User.Name}** angenommen.");
-                if (combi.UserId == (long)Context.User.Id)
+                if (combi.UserId == Context.User.Id)
                     await ReplyAsync($"{Context.User.Mention} du hast erfolgreich die Combi mit **{combi.CombiUser.Name}** angenommen.");
             }
         }
